[C++] Made XCode project and fixed warnings
git-svn-id: https://apps.man.poznan.pl/svn/muscle/muscle/trunk@818 b725f6ce-2418-43e8-b868-1e0ad3d2b257

+++ src/cpp/muscle2/communicator.cpp
@@ -12,7 +12,7 @@
 #include <arpa/inet.h>
 #include <netdb.h>
 
-extern "C" int communicator_read_from_socket(void *socket_handle, void *buf, int buf_len)
+extern "C" size_t communicator_read_from_socket(void *socket_handle, void *buf, int buf_len)
 {
 #ifdef CPPMUSCLE_TRACE
 	cout << "xdr_read_from_socket:" << buf_len << endl;
@@ -20,7 +20,7 @@ extern "C" int communicator_read_from_socket(void *socket_handle, void *buf, int
 	return read(*(int *)socket_handle, buf, buf_len);
 }
 
-extern "C" int communicator_write_to_socket(void *socket_handle, void *buf, int buf_len)
+extern "C" size_t communicator_write_to_socket(void *socket_handle, void *buf, int buf_len)
 {
 #ifdef CPPMUSCLE_TRACE
 	cout << "xdr_write_to_socket:" << buf_len << endl;
